Refactoring to reduce repetition
