+++ src/parser/define.20.ts
@@ -6,10 +6,20 @@
 
 import * as xml2js from "xml2js";
 import { convertAttributeNameToLowerCamelCase, removeNamespaces } from "parser/utils";
+import {
+    parseTranslatedText,
+    parseAliases,
+    parseLeafs,
+    parseWhereClauses,
+    parseEnumeratedItems,
+    parseCodeListItems,
+    parseExternalCodeList,
+    parseFormalExpressions,
+    parseGlobalVariables,
+} from "parser/define.core";
 import type {
     Odm,
     Study,
-    GlobalVariables,
     MetaDataVersion,
     ItemGroupDef,
     ItemGroupDefClassNames,
@@ -18,58 +28,17 @@ import type {
     ItemDefDataType,
     ItemRef,
     ValueListDef,
-    WhereClauseDef,
-    RangeCheck,
-    Comparator,
-    SoftHard,
     CodeList,
     CodeListType,
-    EnumeratedItem,
-    CodeListItem,
-    ExternalCodeList,
     CommentDef,
     MethodDef,
-    FormalExpression,
     Origin,
     OriginType,
-    Leaf,
-    Alias,
     DefineXml,
 } from "interfaces/define.xml.20.d.ts";
 import { ArmDefineXml20 } from "interfaces/arm.10";
 import { parseAnalysisResultDisplays } from "parser/arm.10";
-import { parseTranslatedText, parseDocumentRefs } from "parser/define.20.core";
-
-const parseAliases = (aliasesRaw: any[]): Alias[] => {
-    return aliasesRaw.map((aliasRaw) => ({
-        context: aliasRaw["$"]["context"],
-        name: aliasRaw["$"]["name"],
-    }));
-};
-
-const parseLeafs = (leafsRaw: any[]): { leafs: Record<string, Leaf>; leafsOrder: string[] } => {
-    const leafs: Record<string, Leaf> = {};
-    const leafsOrder: string[] = [];
-
-    // If there are no leafs, return empty object and array
-    if (!leafsRaw) return { leafs, leafsOrder };
-
-    leafsRaw.forEach((leafRaw) => {
-        // If the leaf does not have attributes, skip it
-        if (!leafRaw["$"]) return;
-
-        const leaf: Leaf = {
-            id: leafRaw["$"]["id"],
-            xlink_href: leafRaw["$"]["xlink:href"],
-            title: leafRaw["title"][0],
-        };
-
-        leafs[leaf.id] = leaf;
-        leafsOrder.push(leaf.id);
-    });
-
-    return { leafs, leafsOrder };
-};
+import { parseDocumentRefs } from "parser/define.20.core";
 
 const parseOrigin = (originRaw: any): Origin => {
     const result: Origin = {
@@ -83,109 +52,6 @@ const parseOrigin = (originRaw: any): Origin => {
         result.documentRefs = parseDocumentRefs(originRaw["documentRef"]);
     }
 
-    return result;
-};
-
-const parseCheckValues = (checkValuesRaw: any[]): string[] => {
-    if (!checkValuesRaw) return [];
-    return checkValuesRaw.map((cv) => cv["_"] || cv);
-};
-
-const parseRangeChecks = (rangeChecksRaw: any[]): RangeCheck[] => {
-    if (!rangeChecksRaw) return [];
-    return rangeChecksRaw.map((rc) => ({
-        comparator: rc["$"]["comparator"] as Comparator,
-        softHard: rc["$"]["softHard"] as SoftHard,
-        itemOid: rc["$"]["itemOid"],
-        checkValues: rc["checkValue"] ? parseCheckValues(rc["checkValue"]) : [],
-    }));
-};
-
-const parseWhereClauses = (
-    whereClausesRaw: any[]
-): { whereClauseDefs: Record<string, WhereClauseDef>; whereClauseDefsOrder: string[] } => {
-    const whereClauseDefs: Record<string, WhereClauseDef> = {};
-    const whereClauseDefsOrder: string[] = [];
-    if (!whereClausesRaw) return { whereClauseDefs, whereClauseDefsOrder };
-    whereClausesRaw.forEach((wcRaw) => {
-        if (!wcRaw["$"]) return;
-        const whereClause: WhereClauseDef = {
-            oid: wcRaw["$"]["oid"],
-            rangeChecks: wcRaw["rangeCheck"] ? parseRangeChecks(wcRaw["rangeCheck"]) : [],
-        };
-        if (wcRaw["$"]["commentOid"]) {
-            whereClause.commentOid = wcRaw["$"]["commentOid"];
-        }
-        whereClauseDefs[whereClause.oid] = whereClause;
-        whereClauseDefsOrder.push(whereClause.oid);
-    });
-    return { whereClauseDefs, whereClauseDefsOrder };
-};
-
-const parseEnumeratedItems = (enumeratedItemsRaw: any[]): EnumeratedItem[] => {
-    if (!enumeratedItemsRaw) return [];
-    return enumeratedItemsRaw.map((item) => {
-        const enumeratedItem: EnumeratedItem = {
-            codedValue: item["$"]["codedValue"],
-        };
-        if (item["$"]["rank"]) {
-            enumeratedItem.rank = Number(item["$"]["rank"]);
-        }
-        if (item["$"]["orderNumber"]) {
-            enumeratedItem.orderNumber = Number(item["$"]["orderNumber"]);
-        }
-        if (item["$"]["extendedValue"]) {
-            if (item["$"]["extendedValue"] !== "Yes") {
-                throw new Error(`Invalid value for EnumeratedItem.extendedValue: ${item["$"]["extendedValue"]}`);
-            } else {
-                enumeratedItem.extendedValue = true;
-            }
-        }
-        if (item["alias"]) {
-            enumeratedItem.alias = parseAliases(item["alias"]);
-        }
-        return enumeratedItem;
-    });
-};
-
-const parseCodeListItems = (codeListItemsRaw: any[]): CodeListItem[] => {
-    if (!codeListItemsRaw) return [];
-    return codeListItemsRaw.map((item) => {
-        const codeListItem: CodeListItem = {
-            codedValue: item["$"]["codedValue"],
-            decode: item["decode"] ? item["decode"].map(parseTranslatedText) : [],
-        };
-        if (item["$"]["rank"]) {
-            codeListItem.rank = Number(item["$"]["rank"]);
-        }
-        if (item["$"]["orderNumber"]) {
-            codeListItem.orderNumber = Number(item["$"]["orderNumber"]);
-        }
-        if (item["$"]["extendedValue"]) {
-            if (item["$"]["extendedValue"] !== "Yes") {
-                throw new Error(`Invalid value for EnumeratedItem.extendedValue: ${item["$"]["extendedValue"]}`);
-            } else {
-                codeListItem.extendedValue = true;
-            }
-        }
-        if (item["alias"]) {
-            codeListItem.alias = parseAliases(item["alias"]);
-        }
-        return codeListItem;
-    });
-};
-
-const parseExternalCodeList = (externalCodeListRaw: any): ExternalCodeList => {
-    const result: ExternalCodeList = {
-        dictionary: externalCodeListRaw["$"]["dictionary"],
-        version: externalCodeListRaw["$"]["version"],
-    };
-    if (externalCodeListRaw["$"]["ref"]) {
-        result.ref = externalCodeListRaw["$"]["ref"];
-    }
-    if (externalCodeListRaw["$"]["href"]) {
-        result.href = externalCodeListRaw["$"]["href"];
-    }
     return result;
 };
 
@@ -241,14 +107,6 @@ const parseCommentDefs = (
         commentDefsOrder.push(comment.oid);
     });
     return { commentDefs, commentDefsOrder };
-};
-
-const parseFormalExpressions = (formalExpressionsRaw: any[]): FormalExpression[] => {
-    if (!formalExpressionsRaw) return [];
-    return formalExpressionsRaw.map((fe) => ({
-        context: fe["$"]["context"],
-        value: fe.value,
-    }));
 };
 
 const parseMethodDefs = (methodsRaw: any[]): { methodDefs: Record<string, MethodDef>; methodDefsOrder: string[] } => {
@@ -512,12 +370,6 @@ const parseMetaDataVersion: ParseMetadata = (metadataRaw: any, hasArm?: boolean)
         return result;
     }
 };
-
-const parseGlobalVariables = (globalVariablesRaw: any): GlobalVariables => ({
-    studyName: globalVariablesRaw["studyName"][0],
-    studyDescription: globalVariablesRaw["studyDescription"][0],
-    protocolName: globalVariablesRaw["protocolName"][0],
-});
 
 interface ParseStudy {
     (studyRaw: any, hasArm: true): ArmDefineXml20["odm"]["study"];
